html/semantics/embedded-content-0/media-elements/video_004.htm

<!DOCTYPE HTML>
<html>
  <head>
    <title>HTML5 Media Elements: 'currentTime' is initially zero.</title>
    <meta content="text/html; charset=UTF-8" http-equiv="Content-Type">
    <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
    <link rel="help" href="http://www.w3.org/TR/html5/video.html#current-playback-position" />
    <meta name="assert" content="'currentTime' is initially zero." />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script type="text/javascript" src="/common/media.js"></script>
  </head>
  <body>
<div id='log'></div>

<video id="video0">Your browser does not support media elements.</video>
<script type="text/javascript">
  test(function() {
         assert_equals(document.getElementById("video0").currentTime, 0);
       }, "'currentTime' is initially zero.");
</script>
</div>
</body>
</html>
